third_party/WebKit/LayoutTests/media/track/track-kind.html

<!DOCTYPE html>
<title>Tests that the "kind" attribute is set properly</title>
<video>
    <track src="captions-webvtt/metadata.vtt" default>
</video>
<script src="../media-file.js"></script>
<script src="../media-controls.js"></script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    var trackElement = document.querySelector("track");

    video.oncanplaythrough = t.step_func(testKind);
    trackElement.onload = t.step_func(testKind);
    
    video.src = findMediaFile("video", "../content/counting");

    var eventCount = 0;
    function testKind() {
        eventCount++;
        if (eventCount != 2)
            return;

        // Test default attribute value.
        assert_equals(trackElement.getAttribute("kind"), null);
        assert_equals(trackElement.kind, "subtitles");
        assert_equals(video.textTracks[0].kind, "subtitles");

        // Set to bogus value, should return invalid value default.
        
        // Set via DOM.
        var value = "english-metadata"
        trackElement.removeAttribute("kind");
        trackElement.kind = value;
        assert_equals(trackElement.getAttribute("kind"), value);
        assert_equals(trackElement.kind, "metadata");
        assert_equals(trackElement.track.kind, "metadata");
        
        // Set via attribute.
        trackElement.removeAttribute("kind");
        trackElement.setAttribute("kind", value);
        assert_equals(trackElement.getAttribute("kind"), value);
        assert_equals(trackElement.kind, "metadata");
        assert_equals(trackElement.track.kind, "metadata");

        // Set to known values.
        setKindAndCheck("subtitles");
        setKindAndCheck("captions");
        setKindAndCheck("descriptions");
        setKindAndCheck("chapters");
        setKindAndCheck("metadata");

        t.done();
    }

    function setKindAndCheck(value) {
        // set via DOM.
        trackElement.removeAttribute("kind");
        trackElement.kind = value;
        assert_equals(trackElement.getAttribute("kind"), value);
        assert_equals(trackElement.kind, value);
        assert_equals(trackElement.track.kind, value);
        // Set via attribute.
        trackElement.removeAttribute("kind");
        trackElement.setAttribute("kind", value);
        assert_equals(trackElement.kind, value);
        assert_equals(trackElement.track.kind, value);
        assert_equals(trackElement.getAttribute("kind"), value);
    }
});
</script>
